Extract AlertBase from Alert
