+++ src/Alert/Alert.stories.tsx
@@ -1,7 +1,9 @@
+import Typography from "@material-ui/core/Typography";
 import { Meta, Story } from "@storybook/react";
 import React from "react";
 
 import { Alert, AlertProps } from "./Alert";
+import { AlertBase } from "./AlertBase";
 
 const props: AlertProps = {
   close: false,
@@ -23,6 +25,13 @@ export const WithContentAndClose: Story = () => (
   <Alert {...props} close>
     Lorem ipsum dolor sit amet, consectetur adipiscing elit.
   </Alert>
+);
+export const Custom: Story = () => (
+  <AlertBase {...props} elevation={16} variant="warning">
+    <Typography>
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit.
+    </Typography>
+  </AlertBase>
 );
 
 export default {
