+++ src/Alert/Alert.tsx
@@ -1,5 +1,3 @@
-import Card from "@material-ui/core/Card";
-import CardContent from "@material-ui/core/CardContent";
 import IconButton from "@material-ui/core/IconButton";
 import Typography from "@material-ui/core/Typography";
 import CloseIcon from "@material-ui/icons/Close";
@@ -7,13 +5,11 @@ import clsx from "clsx";
 import React from "react";
 
 import { Error, Info, Success, Warning } from "../icons";
+import { AlertBase, AlertBaseProps, AlertVariant } from "./AlertBase";
 import useStyles from "./styles";
 
-export type AlertVariant = "error" | "warning" | "success" | "info";
-export interface AlertProps {
-  className?: string;
+export interface AlertProps extends AlertBaseProps {
   close?: boolean;
-  variant: AlertVariant;
   title: string;
 }
 const Icon: React.FC<{ variant: AlertVariant }> = ({ variant }) => {
@@ -30,11 +26,11 @@ const Icon: React.FC<{ variant: AlertVariant }> = ({ variant }) => {
 };
 
 export const Alert: React.FC<AlertProps> = ({
-  className,
   close = true,
   variant = "info",
   title,
   children,
+  ...rest
 }) => {
   const classes = useStyles();
   const [visible, setVisible] = React.useState(true);
@@ -44,44 +40,39 @@ export const Alert: React.FC<AlertProps> = ({
   }
 
   return (
-    <Card
-      elevation={0}
-      className={clsx(className, classes.root, {
-        [classes.error]: variant === "error",
-        [classes.warning]: variant === "warning",
-        [classes.success]: variant === "success",
-      })}
-      data-test="alert"
-    >
-      <CardContent>
-        <div className={classes.container}>
-          <div>
-            <Icon variant={variant} />
-          </div>
-          <div className={classes.content}>
-            <div className={classes.titleBar}>
-              <Typography variant="h5">{title}</Typography>
-              {close && (
-                <IconButton
-                  className={clsx(classes.close, {
-                    [classes.closeNoContent]: !!children,
-                  })}
-                  onClick={() => setVisible(false)}
-                  data-test="close"
-                >
-                  <CloseIcon />
-                </IconButton>
-              )}
-            </div>
-            {typeof children === "string" ? (
-              <Typography variant="body1">{children}</Typography>
-            ) : (
-              children
+    <AlertBase variant={variant} {...rest}>
+      <div className={classes.container}>
+        <div>
+          <Icon variant={variant} />
+        </div>
+        <div className={classes.content}>
+          <div className={classes.titleBar}>
+            <Typography variant="h5">{title}</Typography>
+            {close && (
+              <IconButton
+                className={clsx(classes.close, {
+                  [classes.closeNoContent]: !!children,
+                })}
+                onClick={() => setVisible(false)}
+                data-test="close"
+              >
+                <CloseIcon />
+              </IconButton>
             )}
           </div>
+          {typeof children === "string" ? (
+            <Typography variant="body1">{children}</Typography>
+          ) : (
+            children
+          )}
         </div>
-      </CardContent>
-    </Card>
+        {typeof children === "string" ? (
+          <Typography variant="body1">{children}</Typography>
+        ) : (
+          children
+        )}
+      </div>
+    </AlertBase>
   );
 };
