+++ src/Alert/index.ts
@@ -1,1 +1,2 @@
 export * from "./Alert";
+export * from "./AlertBase";
